#16 [REFACTOR] ErrorResponse로 메세지 반환하게 수정
