+++ src/main/java/com/developlife/reviewtwits/handler/MailExceptionHandler.java
@@ -2,14 +2,11 @@ package com.developlife.reviewtwits.handler;
 
 import com.developlife.reviewtwits.exception.mail.MailSendException;
 import com.developlife.reviewtwits.exception.user.VerifyCodeException;
-import com.developlife.reviewtwits.message.response.user.ErrorResponse;
+import com.developlife.reviewtwits.message.response.ErrorResponse;
 import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-import java.util.List;
 
 /**
  * @author ghdic
@@ -19,13 +16,21 @@ import java.util.List;
 public class MailExceptionHandler {
     @ExceptionHandler(MailSendException.class)
     @ResponseStatus(HttpStatus.BAD_REQUEST)
-    public String mailSendExceptionHandler(MailSendException e) {
-        return e.getMessage();
+    public ErrorResponse mailSendExceptionHandler(MailSendException e) {
+        return ErrorResponse.builder()
+                .message(e.getMessage())
+                .errorType(e.getClass().getSimpleName())
+                .fieldName("accountId")
+                .build();
     }
 
     @ExceptionHandler(VerifyCodeException.class)
     @ResponseStatus(HttpStatus.BAD_REQUEST)
-    public String verifyCodeExceptionHandler(VerifyCodeException e) {
-        return e.getMessage();
+    public ErrorResponse verifyCodeExceptionHandler(VerifyCodeException e) {
+        return ErrorResponse.builder()
+                .message(e.getMessage())
+                .errorType(e.getClass().getSimpleName())
+                .fieldName("authenticationCode")
+                .build();
     }
 }
