+++ src/main/java/com/developlife/reviewtwits/handler/ValidExceptionHandler.java
@@ -1,6 +1,6 @@
 package com.developlife.reviewtwits.handler;
 
-import com.developlife.reviewtwits.message.response.user.ErrorResponse;
+import com.developlife.reviewtwits.message.response.ErrorResponse;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
